iwdg動作確認。

+++ src/hardware/circuit/function/delay.cpp
@@ -1,11 +1,11 @@
 #include "delay.hpp"
 
-void delay_us(uint16_t time)
+void delay_us(uint32_t time)
 {
     F405ControlBoard::delay_us(time);
 }
 
-void delay_ms(uint16_t time)
+void delay_ms(uint32_t time)
 {
     F405ControlBoard::delay_ms(time);
 }
