+++ src/hardware/circuit/function/delay.hpp
@@ -1,5 +1,6 @@
 #pragma once
+#include <stdint.h>
 #include "../F405ControlBoard.hpp"
 
-void delay_us(uint16_t time);
-void delay_ms(uint16_t time);
+void delay_us(uint32_t time);
+void delay_ms(uint32_t time);
